Issue #OS-56 Refactoring packages name
