+++ java/registry-interceptor/src/main/java/io/opensaber/registry/interceptor/RDFConversionInterceptor.java
@@ -12,8 +12,9 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.ModelAndView;
 import com.google.gson.Gson;
+
 import io.opensaber.registry.interceptor.handler.BaseRequestHandler;
-import io.opensaber.registry.interceptor.request.transform;
+import io.opensaber.registry.interceptor.request.transform.RequestTransformFactory;
 import io.opensaber.registry.middleware.Middleware;
 import io.opensaber.registry.middleware.MiddlewareHaltException;
 import io.opensaber.registry.middleware.impl.RDFConverter;
