+++ java/registry-interceptor/src/main/java/io/opensaber/registry/interceptor/request/transform/IRequestTransformer.java
@@ -2,8 +2,9 @@ package io.opensaber.registry.interceptor.request.transform;
 
 import java.io.IOException;
 
-import io.opensaber.registry.middleware.transform.commons.Data;
-import io.opensaber.registry.middleware.transform.commons.TransformationException;
+import io.opensaber.registry.middleware.transform.commoms.Data;
+import io.opensaber.registry.middleware.transform.commoms.TransformationException;
+
 
 public interface IRequestTransformer<T> {
